requirements.txt
django-browser-reload
django-tailwind
django-restframework
django-filters
django-corsheaders
